+++ data/beaconrestapi/src/integration-test/java/tech/pegasys/teku/beaconrestapi/v1/debug/GetStateIntegrationTest.java
@@ -22,21 +22,18 @@ import static tech.pegasys.teku.beaconrestapi.RestApiConstants.HEADER_ACCEPT_OCT
 import java.io.IOException;
 import okhttp3.Response;
 import org.apache.tuweni.bytes.Bytes;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import tech.pegasys.teku.api.response.v1.debug.GetStateResponse;
 import tech.pegasys.teku.beaconrestapi.AbstractDataBackedRestAPIIntegrationTest;
 import tech.pegasys.teku.beaconrestapi.handlers.v1.debug.GetState;
+import tech.pegasys.teku.spec.SpecMilestone;
 import tech.pegasys.teku.spec.datastructures.state.beaconstate.BeaconState;
 
 public class GetStateIntegrationTest extends AbstractDataBackedRestAPIIntegrationTest {
-  @BeforeEach
-  public void setup() {
-    startRestAPIAtGenesis();
-  }
 
   @Test
   public void shouldGetStateAsJson() throws IOException {
+    startRestAPIAtGenesis(SpecMilestone.PHASE0);
     final Response response = get("head", HEADER_ACCEPT_JSON);
     assertThat(response.code()).isEqualTo(SC_OK);
     final GetStateResponse stateResponse =
@@ -46,6 +43,7 @@ public class GetStateIntegrationTest extends AbstractDataBackedRestAPIIntegratio
 
   @Test
   public void shouldGetStateAsJsonWithoutHeader() throws IOException {
+    startRestAPIAtGenesis(SpecMilestone.PHASE0);
     final Response response = get("head");
     assertThat(response.code()).isEqualTo(SC_OK);
     final GetStateResponse stateResponse =
@@ -55,6 +53,7 @@ public class GetStateIntegrationTest extends AbstractDataBackedRestAPIIntegratio
 
   @Test
   public void shouldGetStateAsOctetStream() throws IOException {
+    startRestAPIAtGenesis(SpecMilestone.PHASE0);
     final Response response = get("head", HEADER_ACCEPT_OCTET);
     assertThat(response.code()).isEqualTo(SC_OK);
     final BeaconState state =
@@ -66,7 +65,15 @@ public class GetStateIntegrationTest extends AbstractDataBackedRestAPIIntegratio
 
   @Test
   public void shouldRejectUnsupportedAcceptTypes() throws IOException {
+    startRestAPIAtGenesis(SpecMilestone.PHASE0);
     final Response response = get("head", "invalid");
+    assertThat(response.code()).isEqualTo(SC_BAD_REQUEST);
+  }
+
+  @Test
+  public void shouldRejectAltairStateRequestsForJson() throws IOException {
+    startRestAPIAtGenesis(SpecMilestone.ALTAIR);
+    final Response response = get("head");
     assertThat(response.code()).isEqualTo(SC_BAD_REQUEST);
   }