+++ data/serializer/src/main/java/tech/pegasys/teku/provider/JsonProvider.java
@@ -19,6 +19,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.module.SimpleModule;
 import org.apache.tuweni.bytes.Bytes;
 import org.apache.tuweni.bytes.Bytes32;
+import tech.pegasys.teku.api.response.v2.debug.GetStateResponseV2;
 import tech.pegasys.teku.api.response.v2.validator.GetNewBlockResponseV2;
 import tech.pegasys.teku.api.schema.BLSPubKey;
 import tech.pegasys.teku.api.schema.BLSSignature;
@@ -54,6 +55,8 @@ public class JsonProvider {
 
     module.addDeserializer(
         GetNewBlockResponseV2.class, new GetNewBlockResponseV2Deserializer(objectMapper));
+    module.addDeserializer(
+        GetStateResponseV2.class, new GetStateResponseV2Deserializer(objectMapper));
 
     objectMapper.registerModule(module);
   }