+++ CHANGELOG.md
@@ -24,4 +24,5 @@ For information on changes in released versions of Teku, see the [releases page]
 - implement POST `/eth/v1/validator/duties/sync/{epoch}` for Altair fork.
 - implement GET and POST `/eth/v1/validator/sync_committee_subscriptions` for Altair fork.
 - implement GET `/eth/v2/validator/blocks/{slot}` for Altair fork.
+- implement GET `/eth/v2/debug/beacon/states/:state_id` for Altair fork.
 - `/eth/v1/validator/blocks/{slot}` will now produce an altair block if an altair slot is requested.